Makes some changes to file structure and tries to make more API call work - it doesn't
